Time: 19 ms (5.31%), Space: 45.4 MB (89.63%) - LeetHub

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.java
@@ -52,7 +52,7 @@ class Solution {
         for(int i=1;i<=h;i++){
             double sum=solve(root, i);
             int count=countNodesAtLevel(root,i);
-            ans.add(sum/(double)count);
+            ans.add(sum/count);
         }
         return ans;
     }
